64 bit stream offsets and 64 bit PS integers objects
Revise the stream and "gp_" file APIs to support file sizes >4Gb (on suitable
systems).

I've added a gs_offset_t type which, by default is a 64 bit signed type (this
matches the standard lib's 64 bit "off_t" type). For ports that may want to
use an alternative type, define "GS_OFFSET_T" to the desired type on the
compiler command line.

By default, Ghostscript will now use 64 bit integer objects in Postscript. Again,
for porting ease, one can opt for 32 bit integer objects by defining PSINT32BIT=1
on the compiler command line.

The handling of integer objects in various operators has been "augmented" so that
when working in "CPSI compatibility mode", intepreting and manipulating integer
objects will see them "clamped" to 32 bit signed values. This allows Ghostscript
to still pass the Quality Logic tests which spuriously rely on 32 bit integer
objects.

There is a minor tweek in the PS interpreter's handling of Truetype fonts (*not*
Type 42), where the algorithm used to create an XUID relied on 32 bit integers.

There is a small performance penalty in this, but my testing suggests that it is
smaller than the time resolution of our testing capabilities (i.e. there are
greater variances between test runs than between the 32 and 64 bit objects code).

We can now interpret PDF files >4Gb.

In addition, the graphics library will now include inttypes.h (in stdint_.h) if
it is available in order to make the "PRI...." printf and co formatting macros
available. If inttypes.h is not available, there are fallback definitions of the
most used of those. These can also be defined on the compiler command line for
specific ports, if required.

A custom "PRI..." macro has been added in the Postscript interpreter,
"PRIpsint", which is defined appropriately when integer objects are 64 or
32 bit.

This commit does *not* revise the device API to handle 64 bit integer values
(in the one case where this is required, we use a string to hold the 64 bit value,
I suggest continuing this, so the device API remains consistent across ports).

I do expect a bug tail with this, as we start to encounter more >4Gb PDFs.

Cluster differences are restricted to the 3 QL Postscript test jobs that have
ended up in comparefiles (and thus are not run with the gs_cet.ps setup).

+++ gs/base/gdevpsf2.c
@@ -1731,7 +1731,7 @@ psf_write_cid0_font(stream *s, gs_font_cid0 *pfont, int options,
                         pfont->cidata.FDArray[0], true);
 
     /* Write the charset. */
-    if_debug1m('l', s->memory, "[l]charset = %lu\n", stell(writer.strm) - start_pos);
+    if_debug1m('l', s->memory, "[l]charset = %"PRId64"\n", (int64_t)(stell(writer.strm) - start_pos));
     cff_write_cidset(&writer, &genum);
 
     /* Write the FDSelect structure, checking the offset. */
